go_pybindings

Module analysis

Functions

py_calc_edep_simple

py_cfd_simple

Find the peak onset time based on a cfd “Constant fraction discrimination” algorithm

py_find_peaks

The GAPS peak finding algorithm, based on legacy code written by the UCLA TOF team.

py_find_peaks_zscore

py_find_zero_crossings

Get a vector with the indizes where the input array crosses zero

py_get_periods

py_integrate

py_interpolate_time

py_time2bin