go_pybindings

Module caraspace

The following file is part of gaps-online-software and published under the GPLv3 license

This file contains the source for pybindings with pyO3 for the caraspace i/o system

Structs

PyCRFrame

The building blocks of the caraspace serialization library

PyCRFrameObject

The building blocks of the caraspace serialization library

PyCRReader

Read caraspace files. Caraspace files are an aggregate filetype which allows to hold information from multiple sources in an efficient binary format. For the use within the GAPS experiment, L0 files are caraspace files and contain the TOF waveforms as sourced by the files written by the TOFComputer to disk as well as the files emitted by the flight computer (telemetry).

PyCRWriter

Functions

py_parse_u8

Parse an u8 from python bytes.

py_parse_u16

Parse an u16 from python bytes.

py_parse_u32

Parse an u32 from python bytes.

py_parse_u64

Parse an u64 from python bytes.