go_pybindings

Module command_factory

The wrapped commands from the command factory

Functions

py_change_datapublisherconfig

Change the data publisher config part of the config file

py_change_tofeventbuilderconfig

Change the TOFEventBuilderSettings in the config

py_change_tofrbconfig

Change the RB config part of the config file

py_change_tofrunconfig

Change the run config part of the config file

py_change_triggerconfig

Change the MTBSettings in the config file with relevant trigger settings

py_get_ratpdumap_hardcoded

A hardwired map of PDU #id PDUCHANNEL #id to (RAT,RAT)

py_get_ratrbmap_hardcoded

A hardwired map of RAT -> (RB1, RB2)

py_get_rbratmap_hardcoded

A hardwired map of RB -> RAT

py_rb_calibration

Run a calibration of all RBs

py_restart_liftofrb

Restart the liftof-rb clients on the given boards

py_shutdown_all_rbs

Send the ‘sudo shutdown now’ command to ALL RBs

py_shutdown_rat

Send the ‘sudo shutdown now’ command to all RBs in a RAT

py_shutdown_ratpair

Send the ‘sudo shutdown now’ command to all RBs in the 2 RATs connected to a certain PDU channel

py_shutdown_rb

Send the ‘sudo shutdown now’ command to a single RB

py_shutdown_tofcpu

Send the ’sudo shutdown now command to the TOF main computer (“TOFCPU”)

py_start_run

Trigger the start of a new data run with the next active config

py_stop_run

Stop the current active run and idle